Extracts 'identify' as a seperate service
 - also removes hostname response, ya know... for those nosy coworkers :P
