+++ bot.js
@@ -10,7 +10,6 @@ if (!process.env.token) {
 }
 
 var Botkit = require('./lib/Botkit.js')
-var os = require('os')
 
 var controller = Botkit.slackbot({
   debug: true
@@ -22,30 +21,4 @@ controller.spawn({
 
 controller.hears([constants.LOOKUP], constants.ADDRESSED, services.stackOverflow)
 controller.hears([constants.DEFINE], constants.ADDRESSED, services.urbanDictionary)
-
-controller.hears(['uptime'], constants.ADDRESSED, function (bot, message) {
-  var hostname = os.hostname()
-  var uptime = formatUptime(process.uptime())
-
-  bot.reply(message, ':robot_face: I am a bot named <@' + bot.identity.name + '>. I have been running for ' + uptime + ' on ' + hostname + '.')
-})
-
-function formatUptime (uptime) {
-  var unit = 'second'
-  if (uptime > 60) {
-    uptime = uptime / 60
-    unit = 'minute'
-  }
-
-  if (uptime > 60) {
-    uptime = uptime / 60
-    unit = 'hour'
-  }
-
-  if (uptime !== 1) {
-    unit = unit + 's'
-  }
-
-  uptime = uptime + ' ' + unit
-  return uptime
-}
+controller.hears([constants.IDENTIFY], constants.ADDRESSED, services.identify)
