+++ services/identity.js
@@ -1,0 +1,32 @@
+var request = require('request')
+var constants = require('../constants')
+var os = require('os')
+
+module.exports = function (bot, message, cb) {
+
+  var hostname = os.hostname()
+  var uptime = formatUptime(process.uptime())
+
+  bot.reply(message, ':robot_face: I am a bot named <@' + bot.identity.name + '>. I have been running for ' + uptime)
+}
+
+
+function formatUptime (uptime) {
+  var unit = 'second'
+  if (uptime > 60) {
+    uptime = uptime / 60
+    unit = 'minute'
+  }
+
+  if (uptime > 60) {
+    uptime = uptime / 60
+    unit = 'hour'
+  }
+
+  if (uptime !== 1) {
+    unit = unit + 's'
+  }
+
+  uptime = uptime + ' ' + unit
+  return uptime
+}
