+++ services/index.js
@@ -1,7 +1,9 @@
 var stackOverflow = require('./stack-overflow')
 var urbanDictionary = require('./urban-dictionary')
+var identify = require('./identity')
 
 module.exports = {
   stackOverflow,
-  urbanDictionary
+  urbanDictionary,
+  identify
 }
